Cleanup files and Code
Cleaned up project files:
New structure into Code and Plot folders. Old Code and files go into the Archive folder

Added function: plot_distribution_histograms
Input a dataset with the three trust variables and an index ,and it will save the individual and stacked distributions into the graph folder. (so we can also do it for the seperate regions.